+++ Gui/TimeLineGui.h
@@ -124,6 +124,8 @@ public:
 
     void setFrameRangeEdited(bool edited);
 
+    void setTimeFormatFrames(bool value);
+
     void centerOn_tripleSync(SequenceTime left, SequenceTime right);
 
     void getVisibleRange(SequenceTime* left, SequenceTime* right) const;
@@ -171,6 +173,7 @@ public Q_SLOTS:
 
     void onProjectFrameRangeChanged(int, int);
 
+    void onTimeFormatChanged(int);
 
 
 private: